Use the new dynamic_objectt class for dynamic object splitting.
Instances of a dynamic object are now stored along with their guards.
The resulting expression can be generated from these as a case split where the
correct instance is chosen if its allocation guard holds.
The order of instances is important, first guards will be checked first.
Different guards can be generated for concrete instances and for
non-deterministic instances.