Introduce analysis of used array indices
The analysis collects all expressions that are used as array indices in
the program.

Written indices (those occurring at LHS) are used for array partitioning
in the array abstract domain. These are collected per-location - for
each location and each array, the analysis collects all indices of the
array that might have been assigned to before in the program.

The computed invariants are then projected onto the read indices (those
occurring at RHS). These are collected globally for the whole program -
for each array, the analysis collects all indices of the array that are
read from somewhere in the program.

In the array domain, we suppose that each array has a single written
index.

Add test that uses more than one read index.