Arrays: improve invariant projection
Instead of projecting the whole invariant onto each read index, which is
quite inefficient, only add projection of symbolic segment indices onto
read indices.

For each segment of an array and for each index read from that array,
create an implication:
    (idx#read >= lower && idx#read < upper) => idx#read == idx#segment

Also, the projection must be used during invariant inference in order to
be able to use already computed parts of the invariant when computing
other parts (e.g. to use an invariant of one loop to correctly compute
an invariant of the following loop). This is ensured by adding the
projection to pre-constraint.

An example that requires this extension is added to the regression
tests.