Arrays: enable arbitrary domain for array elements
The array domain is transformed from being a simple (row-based) domain
into being a general domain that has an inner domain for computing
invariants for array elements.

All elements in an array segment are represented by a fresh symbolic
variable of an appropriate type that is used to compute an invariant for
the elements.

The array strategy solver runs a solver for the inner domain in a
context of bindings between symbolic element variables (used in the
invariant) and actually read array elements (used in the program).

For now, the interval domain is always used as the inner domain.

A test requiring an interval invariant for array elements is added.