Introduce expression dependence analysis
The analysis determines which expressions have a potential relation
between them by checking which expressions occur in the same
expression/statement. More specifically, it creates an equivalence
partition based on the relation "occur in the same expression".

This information can be used to make the abstract domain template
smaller. Currently, it is used for array domain combined with zones
domain.