Arrays: compute differences with scalars
This enables zones domain with arrays. If the zones domain is chosen,
some custom differences are added.

In particular, this commit adds difference between each array segment
and all dependent scalar values (both when the scalars are updated
within the same loop and when they're not).

Allows to verify, e.g., computing an array maximum (if the maximum is
computed in the same loop) or array initialization from a variable
value. Add regression tests for such programs.